Added Result wrapper class to provide funtionality for error responses. Added logic to BaseAppController  to check Result objects for errors. Wrapped API endpoint methods in function from BaseAppController to check errors.

+++ Application/Parks/Delete.cs
@@ -1,4 +1,5 @@
-﻿using Application.Core.Interfaces;
+﻿using Application.Core;
+using Application.Core.Interfaces;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -12,25 +13,28 @@ namespace Application.Parks
     public class Delete
     {
 
-        public class Command : IRequest
+        public class Command : IRequest<Result<Unit>>
         {
             public Guid Id { get; set; }
         }
 
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command, Result<Unit>>
         {
             private readonly IApplicationDbContext _db;
             public Handler(IApplicationDbContext db)
             {
                 _db = db;
             }
-            public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var park = await _db.Parks.FindAsync(request.Id);
+                if (park == null) return null;
 
                 _db.Parks.Remove(park);
-                await _db.SaveChangesAsync();
-                return Unit.Value;
+
+                var result = await _db.SaveChangesAsync() > 0;
+                if (!result) return Result<Unit>.Failure("Failed to delete Park");
+                return Result<Unit>.Success(Unit.Value);
             }
         }
     }
